Consistent-ize logging calls.
