+++ src/Microsoft.EntityFrameworkCore.Relational/Query/ExpressionTranslators/EqualsTranslator.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Linq.Expressions;
 using JetBrains.Annotations;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
 using Microsoft.Extensions.Logging;
@@ -46,8 +47,9 @@ namespace Microsoft.EntityFrameworkCore.Query.ExpressionTranslators
                             Expression.Convert(argument, unwrappedArgumentType));
                     }
 
-                    _logger.LogInformation(
-                        RelationalStrings.PossibleUnintendedUseOfEquals(
+                    _logger.LogWarning(
+                        RelationalLoggingEventId.PossibleUnintendedUseOfEquals,
+                        () => RelationalStrings.PossibleUnintendedUseOfEquals(
                             methodCallExpression.Object.ToString(),
                             argument.ToString()));
