+++ src/Microsoft.EntityFrameworkCore/Extensions/Internal/CoreLoggerExtensions.cs
@@ -3,10 +3,9 @@
 
 using System;
 using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.Extensions.Logging;
 
 // ReSharper disable once CheckNamespace
-namespace Microsoft.EntityFrameworkCore.Internal
+namespace Microsoft.Extensions.Logging
 {
     internal static class CoreLoggerExtensions
     {
