+++ src/Microsoft.EntityFrameworkCore/Infrastructure/CoreLoggingEventId.cs
@@ -39,5 +39,16 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
         ///     A query specified an Include operation that was ignored because the included navigation was not reachable in the final query result.
         /// </summary>
         IncludeIgnored
+        QueryPlan,
+
+        /// <summary>
+        ///     A warning that sensitive data logging is enabled.
+        /// </summary>
+        SensitiveDataLoggingEnabled,
+
+        /// <summary>
+        ///     A warning during model validation.
+        /// </summary>
+        ModelValidationWarning
     }
 }
